Update Layout and Code Splitting plus Lazy loads
