+++ backend/routes/apiRoutes.js
@@ -3,16 +3,19 @@ let router = express.Router();
 const axios = require("axios");
 const API_URL = "https://api.rawg.io/api/";
 const params = {
-  key: "3f4a034d7b034f7bbea4371034a6e66d",
+  key: "3f4a034d7b034f7bbea4371034a6e66d"
 };
 
 
-//GAME DETAIL
-router.get("/:category/:id", (req, res) => {
-  const id = req.params.id;
-
+//!GET THE FOLLOWING API TAG,PLATFORMS ETC
+router.get("/:resource", (req, res) => {
+  const resource = req.params.resource;
+  const params = {
+    key: "3f4a034d7b034f7bbea4371034a6e66d",
+    ordering: req.query.ordering
+  };
   axios
-    .get(`${API_URL}games/${id}`, { params })
+    .get(`${API_URL}${resource}`, { params })
     .then((response) => {
       res.send(response.data);
     })
@@ -20,6 +23,52 @@ router.get("/:category/:id", (req, res) => {
       res.send(error);
     });
 });
+
+//! Game Details
+router.get("/games/:slug", (req, res) => {
+  const slug = req.params.slug;
+  axios
+    .get(`${API_URL}games/${slug}`, { params })
+    .then((response) => {
+      res.send(response.data);
+    })
+    .catch((error) => {
+      res.send(error);
+    });
+});
+
+//! Category List
+
+router.get("/category/:category/:slug", (req, res) => {
+  const category = req.params.category;
+  const slug = req.params.slug;
+  axios
+    .get(`${API_URL}/${category}/${slug}`, { params })
+    .then((response) => {
+      res.send(response.data);
+    })
+    .catch((error) => {
+      res.send(error);
+    });
+});
+
+//! Get Data
+
+router.get("/fetch/:tags/:id", (req, res) => {
+  const tags = req.params.tags;
+  const id = req.params.id;
+  axios.get(`${API_URL}${tags}/${id}`, { params })
+    .then(response => {
+      const games = response.data.results;
+      for (const game of games) {
+        console.log(game.name);
+      }
+    })
+    .catch(error => {
+      console.error('Error fetching games:', error);
+    });
+});
+
 
 //GET GAME SCREENSHOTS
 router.get("/games/:id/screenshots", (req, res) => {
@@ -34,35 +83,7 @@ router.get("/games/:id/screenshots", (req, res) => {
     });
 });
 
-//!GET THE FOLLOWING API TAG,PLATFORMS ETC
-router.get("/:resource", (req, res) => {
-  const resource = req.params.resource;
-  const params = {
-    key: "3f4a034d7b034f7bbea4371034a6e66d",
-    ordering : req.query.ordering
-  };
-  axios
-    .get(`${API_URL}${resource}`, { params })
-    .then((response) => {
-      res.send(response.data);
-    })
-    .catch((error) => {
-      res.send(error);
-    });
-});
 
-// ///!GET THE FOLLOWING API TAG,PLATFORMS ETC
-// router.get("/ddd/:resource/:id", (req, res) => {
-//   const resource = req.params.resource;
-//   axios
-//     .get(`${API_URL}${resource}/${req.params.id}`, { params })
-//     .then((response) => {
-//       res.send(response.data);
-//     })
-//     .catch((error) => {
-//       res.send(error);
-//     });
-// });
 
 
 module.exports = router;
