+++ frontend/src/components/Home.jsx
@@ -1,11 +1,7 @@
-import React from "react";
+import React from 'react'
 
 export default function Home() {
   return (
-    <>
-      <div className="container-fluid" style={{ "height": "100vh" }}>
-        <h1>Welcome to Vapour</h1>
-      </div>
-    </>
-  );
+    <div>Home</div>
+  )
 }
