+++ frontend/src/components/Publishers.jsx
@@ -5,17 +5,17 @@ import Lottie from "lottie-react";
 import Gameloader from "../assets/Lottie/gamecontroller.json";
 import Category_Cards from "./Category_Cards";
 
-export default function Stores() {
-    const [stores, setStores] = useState([]);
+export default function Publishers() {
+    const [publishers, setPublishers] = useState([]);
     const [loading, setLoading] = useState(true);
     const [nextPage, setNextPage] = useState("");
 
     const FetchData = () => {
         setLoading(true)
         axios
-            .get(`http://localhost:8000/api/stores`)
+            .get(`http://localhost:8000/api/publishers`)
             .then((res) => {
-                setStores(res.data.results);
+                setPublishers(res.data.results);
                 setNextPage(res.data.next);
                 setLoading(false);
             })
@@ -35,7 +35,7 @@ export default function Stores() {
                     fetch(nextPage)
                         .then((response) => response.json())
                         .then((data) => {
-                            setStores((prevGames) => [...prevGames, ...data.results]);
+                            setPublishers((prevGames) => [...prevGames, ...data.results]);
                             setNextPage(data.next);
                         })
                         .catch((error) => console.error(error));
@@ -47,12 +47,12 @@ export default function Stores() {
         return () => {
             window.removeEventListener("scroll", handleScroll);
         };
-    }, [stores, nextPage]);
+    }, [publishers, nextPage]);
 
     return (
         <>
             <div className="container-fluid">
-                <h1>Stores</h1>
+                <h1>Publishers</h1>
                 <div className="row">
                     {
                         (loading) ?
@@ -62,9 +62,9 @@ export default function Stores() {
                                 </div>
                             </div>
                             :
-                            stores?.map((item, id) => (
+                            publishers?.map((item, id) => (
                                 <div className="col-6 col-lg-3 my-2" key={id}>
-                                    <Category_Cards item={item} category={'stores'}/>
+                                    <Category_Cards item={item} category={'publishers'} />
                                 </div>
                             ))}
                 </div>
