+++ frontend/src/components/Screenshots.jsx
@@ -1,14 +1,13 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { Carousel } from "react-bootstrap";
-import { GAME_NODE_ROOT } from "../helpers/Api";
 
 export default function Screenshots({ gameid }) {
   const [screenshots, setscreenshots] = useState({});
 
   useEffect(() => {
     axios
-      .get(`${GAME_NODE_ROOT}/games/${gameid}/screenshots`)
+      .get(`http://localhost:8000/api/games/${gameid}/screenshots`)
       .then((res) => {
         setscreenshots(res.data);
       })
