+++ frontend/src/components/Tags.jsx
@@ -3,72 +3,71 @@ import { Link } from "react-router-dom";
 import axios from "axios";
 import Lottie from "lottie-react";
 import Gameloader from "../assets/Lottie/gamecontroller.json";
-import { GAME_NODE_ROOT } from "../helpers/Api";
-import CategoryCard from "./CategoryCard";
+import Category_Cards from "./Category_Cards";
 
 export default function Tags() {
-  const [tags, setTags] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [nextPage, setNextPage] = useState("");
-
-  useEffect(() => {
-    axios
-      .get(`${GAME_NODE_ROOT}/tags`)
-      .then((res) => {
-        setTags(res.data.results);
-        setNextPage(res.data.next);
-        setLoading(false);
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  }, []); //!Initialize first 20 tags
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (window.innerHeight + window.scrollY >= document.body.offsetHeight) {
-        if (nextPage !== null) {
-          fetch(nextPage)
-            .then((response) => response.json())
-            .then((data) => {
-              setTags((prevGames) => [...prevGames, ...data.results]);
-              setNextPage(data.next);
+    const [tags, setTags] = useState([]);
+    const [loading, setLoading] = useState(true);
+    const [nextPage, setNextPage] = useState("");
+    const FetchData = () => {
+        setLoading(true)
+        axios
+            .get(`http://localhost:8000/api/tags`)
+            .then((res) => {
+                console.log(res.data);
+                setTags(res.data.results);
+                setNextPage(res.data.next);
+                setLoading(false);
             })
-            .catch((error) => console.error(error));
-        }
-      }
+            .catch((error) => {
+                console.log(error);
+            });
     };
+    useEffect(() => {
+        FetchData()
+    }, []); //!Initialize first 20 Games
 
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [tags, nextPage]);
+    useEffect(() => {
+        const handleScroll = () => {
+            if (window.innerHeight + window.scrollY >= document.body.offsetHeight) {
+                if (nextPage !== null) {
+                    fetch(nextPage)
+                        .then((response) => response.json())
+                        .then((data) => {
+                            setTags((prevGames) => [...prevGames, ...data.results]);
+                            setNextPage(data.next);
+                        })
+                        .catch((error) => console.error(error));
+                }
+            }
+        };
 
-
-  if (loading) {
+        window.addEventListener("scroll", handleScroll);
+        return () => {
+            window.removeEventListener("scroll", handleScroll);
+        };
+    }, [tags, nextPage]);
     return (
-      <>
-        <div className="d-flex align-items-center justify-content-center vh-100">
-          <div style={{ width: 400 }}>
-            <Lottie animationData={Gameloader} loop={true} />
-          </div>
-        </div>
-      </>
-    );
-  }
-  return (
-    <>
-      <div className="container-fluid">
-        <h1 className="text-white">All Tags</h1>
-        <div className="row">
-          {tags?.map((item, id) => (
-            <div className="col-6 col-lg-3 my-2" key={id}>
-              <CategoryCard item={item} />
+        <>
+            <div className="container-fluid">
+                <h1>Tags</h1>
+
+                <div className="row">
+                    {
+                        (loading) ?
+                            <div className="d-flex align-items-center justify-content-center vh-100">
+                                <div style={{ width: 400 }}>
+                                    <Lottie animationData={Gameloader} loop={true} />
+                                </div>
+                            </div>
+                            :
+                            tags?.map((item, id) => (
+                                <div className="col-6 col-lg-3 my-2" key={id}>
+                                    <Category_Cards item={item} category={'tags'} />
+                                </div>
+                            ))}
+                </div>
             </div>
-          ))}
-        </div>
-      </div>
-    </>
-  );
+        </>
+    );
 }
