--- frontend/src/containers/Footer.jsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-function Footer() {
-  return (
-    <>
-      <footer className="footer mt-auto py-3 bg-body-tertiary">
-        <div className="container">
-          <span className="text-body-secondary">
-            Footer Area Center
-          </span>
-        </div>
-      </footer>
-    </>
-  );
-}
-
-export default Footer;
