--- frontend/src/containers/index.jsx
@@ -1,3 +1,0 @@
-export { default as Footer } from './Footer'
-export { default as ContentArea } from './ContentArea'
-export { default as Header } from './Header'
