+++ frontend/src/index.css
@@ -14,14 +14,11 @@ body {
   background-color: #151515;
 }
 
-@keyframes flip {
-  from {
-    transform: rotateY(0deg);
-  }
-  to {
-    transform: rotateY(180deg);
-  }
+
+#root h1 {
+  color: aliceblue
 }
+
 
 section {
   padding-block: 2rem;
@@ -30,11 +27,14 @@ section {
 .flip-container {
   width: 100%;
   height: 350px;
-  .card {
-    width: 100%;
-    height: 100%;
-  }
+  perspective: 1000px;
 }
+
+.flip-container .card {
+  width: 100%;
+  height: 100%;
+}
+
 .flip-front {
   position: relative;
 }
@@ -46,47 +46,76 @@ section {
   position: absolute;
   border-radius: 0.376rem;
   inset: 0;
-  background-color: rgba(13 11 44 / 50%);
+  background-color: rgba(13, 11, 44, 50%);
   z-index: -999;
 }
 
 .flip-back {
   background: linear-gradient(135deg, #65799b 0%, #5e2563 100%);
 }
-.flip-container {
-  perspective: 1000px;
 
-  &:hover .flip {
+.flip-container:hover .flip {
+  transform: rotateY(180deg);
+}
+
+.flip {
+  width: 100%;
+  height: 100%;
+  transition: 0.6s;
+  transform-style: preserve-3d;
+  position: relative;
+}
+
+.flip-front {
+  z-index: 2;
+  transform: rotateY(0deg);
+}
+
+.flip-back {
+  transform: rotateY(180deg);
+}
+
+.flip-front,
+.flip-back {
+  backface-visibility: hidden;
+  position: absolute;
+  top: 0;
+  left: 0;
+  will-change: transform;
+}
+
+.card-category {
+  position: relative;
+  height: 10rem;
+  z-index: 1;
+}
+
+
+.card-category:hover {
+  cursor: pointer;
+  background-color: rgba(13, 11, 44, 80%);
+}
+
+.card-category::after {
+  content: "";
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  border-radius: 0.376rem;
+  background-color: rgba(13, 11, 44, 50%);
+  inset: 0;
+  z-index: -99999;
+}
+
+
+
+@keyframes flip {
+  from {
+    transform: rotateY(0deg);
+  }
+
+  to {
     transform: rotateY(180deg);
-  }
-
-  .flip {
-    width: 100%;
-    height: 100%;
-  }
-
-  .flip {
-    transition: 0.6s;
-    transform-style: preserve-3d;
-    position: relative;
-
-    .flip-front {
-      z-index: 2;
-      transform: rotateY(0deg);
-    }
-
-    .flip-back {
-      transform: rotateY(180deg);
-    }
-
-    .flip-front,
-    .flip-back {
-      backface-visibility: hidden;
-      position: absolute;
-      top: 0;
-      left: 0;
-      will-change: transform;
-    }
   }
 }
 
@@ -95,28 +124,32 @@ section {
   transition: background-color 0.3s ease-in-out;
   transition-delay: 0.2457s;
 }
-.details__text{
+
+.details__text {
   font-size: 11px;
 }
 
 .maskedbg {
   background-color: rgb(0 0 0 / 60%);
 }
+
 .gametags {
   margin: 2px;
 }
+
 .gametitle h1 {
   font-size: 3.5rem;
 }
+
 /* Set the background image */
 .details {
   background-size: cover;
   color: azure;
   width: 100%;
-  /* height: 100vh; */
   position: relative;
   z-index: 0;
 }
+
 /* Add the darkened effect to the image */
 .details::after {
   content: "";
